+++ Panther/src/sema_to_pir/SemaToPIR.hpp
@@ -322,6 +322,7 @@ namespace pcit::panther{
 
 			[[nodiscard]] auto get_unmangled_func_name(const sema::Func& func) const -> std::string;
 			[[nodiscard]] auto get_unmangled_struct_name(BaseType::Struct::ID struct_id) const -> std::string;
+			[[nodiscard]] auto get_unmangled_union_name(BaseType::Union::ID union_id) const -> std::string;
 
 			
 			template<bool PIR_STMT_NAME_SAFE = false>